fusion/requirements.txt
asgiref==3.5.0
click==8.1.3
colorama==0.4.4
coverage==6.3.3
dj-static==0.0.6
Django==4.0.4
django-stdimage==5.3.0
gunicorn==20.1.0
joblib==1.1.0
model-mommy==2.0.0
nltk==3.7
Pillow==9.1.0
psycopg2-binary==2.9.3
regex==2022.4.24
sqlparse==0.4.2
static3==0.7.0
textblob==0.17.1
tqdm==4.64.0
tzdata==2022.1
